refactor: move processCowFertilizerProduction to its own file
