+++ src/game-logic/reducers/index.test.js
@@ -1,7 +1,6 @@
 import { shapeOf, testCrop, testItem } from '../../test-utils'
 import {
   ACHIEVEMENT_COMPLETED,
-  FERTILIZERS_PRODUCED,
   LOAN_INCREASED,
   LOAN_PAYOFF,
   PRICE_CRASH,
@@ -10,7 +9,6 @@ import {
 import {
   COW_GESTATION_PERIOD_DAYS,
   COW_HUG_BENEFIT,
-  COW_FERTILIZER_PRODUCTION_RATE_SLOWEST,
   MAX_ANIMAL_NAME_LENGTH,
   MAX_DAILY_COW_HUG_BENEFITS,
   MAX_LATEST_PEER_MESSAGES,
@@ -26,7 +24,6 @@ import { genders, standardCowColors } from '../../enums'
 import {
   generateCow,
   getCostOfNextStorageExpansion,
-  getCowFertilizerItem,
   getCowValue,
   getPlotContentFromItemId,
   getPriceEventForCrop,
@@ -395,114 +392,6 @@ describe('computeStateForNextDay', () => {
     expect(firstRow[0].daysWatered).toBe(1)
     expect(firstRow[0].daysOld).toBe(1)
     expect(todaysNotifications).toBeEmpty()
-  })
-})
-
-describe('processCowFertilizerProduction', () => {
-  let state
-
-  beforeEach(() => {
-    state = {
-      cowInventory: [],
-      inventory: [],
-      inventoryLimit: -1,
-      newDayNotifications: [],
-    }
-  })
-
-  describe('cow should not produce fertilizer', () => {
-    test('cow does not produce fertilizer', () => {
-      const baseDaysSinceProducingFertilizer = 2
-
-      state.cowInventory = [
-        generateCow({
-          daysSinceProducingFertilizer: baseDaysSinceProducingFertilizer,
-          gender: genders.MALE,
-        }),
-      ]
-
-      const {
-        cowInventory: [{ daysSinceProducingFertilizer }],
-        inventory,
-        newDayNotifications,
-      } = fn.processCowFertilizerProduction(state)
-
-      expect(daysSinceProducingFertilizer).toEqual(
-        baseDaysSinceProducingFertilizer
-      )
-      expect(inventory).toEqual([])
-      expect(newDayNotifications).toEqual([])
-    })
-  })
-
-  describe('cow should produce fertilizer', () => {
-    describe('inventory space is available', () => {
-      test('cow produces fertilizer and fertilizer is added to inventory', () => {
-        state.cowInventory = [
-          generateCow({
-            color: standardCowColors.WHITE,
-            daysSinceProducingFertilizer: COW_FERTILIZER_PRODUCTION_RATE_SLOWEST,
-            gender: genders.MALE,
-          }),
-        ]
-
-        const {
-          cowInventory: [cow],
-          inventory,
-          newDayNotifications,
-        } = fn.processCowFertilizerProduction(state)
-
-        const { daysSinceProducingFertilizer } = cow
-
-        expect(daysSinceProducingFertilizer).toEqual(0)
-        expect(inventory).toEqual([{ id: 'fertilizer', quantity: 1 }])
-        expect(newDayNotifications).toEqual([
-          {
-            message: FERTILIZERS_PRODUCED`${{
-              [getCowFertilizerItem(cow).name]: 1,
-            }}`,
-            severity: 'success',
-          },
-        ])
-      })
-    })
-
-    describe('inventory space is not available', () => {
-      test('cow produces fertilizer but fertilizer is not added to inventory', () => {
-        state.inventoryLimit = 1
-        state.cowInventory = [
-          generateCow({
-            color: standardCowColors.WHITE,
-            daysSinceProducingFertilizer: COW_FERTILIZER_PRODUCTION_RATE_SLOWEST,
-            gender: genders.MALE,
-          }),
-          generateCow({
-            color: standardCowColors.WHITE,
-            daysSinceProducingFertilizer: COW_FERTILIZER_PRODUCTION_RATE_SLOWEST,
-            gender: genders.MALE,
-          }),
-        ]
-
-        const {
-          cowInventory: [cow],
-          inventory,
-          newDayNotifications,
-        } = fn.processCowFertilizerProduction(state)
-
-        const { daysSinceProducingFertilizer } = cow
-
-        expect(daysSinceProducingFertilizer).toEqual(0)
-        expect(inventory).toEqual([{ id: 'fertilizer', quantity: 1 }])
-        expect(newDayNotifications).toEqual([
-          {
-            message: FERTILIZERS_PRODUCED`${{
-              [getCowFertilizerItem(cow).name]: 1,
-            }}`,
-            severity: 'success',
-          },
-        ])
-      })
-    })
   })
 })
